Builder setter methods shouldn't take optional args
The typical use of builder is one-time initialization (often in a single
statement) so if user doesn't want to set an optional parameter, they
just don't call the relevant setter method. In a rare situation where
this is desirable, user can always create a new config based on an
existing one.