Make it mandatory to pass VM ID to the config::Builder
We still allow user to pass `None` though to specify random ID
generation. This goes against the spirit of the previous changes in this
series where we made all builder setter args mandatory but we need VM
ID to be set in the very beginning and only once, since auxiliary
configuration builders make use of this ID and things will break if user
changes it later on.